tree

+++ ProSoft.UI/Areas/Treasury/Controllers/AccSafeCashController.cs
@@ -130,6 +130,11 @@ namespace ProSoft.UI.Areas.Treasury.Controllers
             ViewBag.userCode = userCode;
             ViewBag.branchId = int.Parse(User.Claims.FirstOrDefault(c => c.Type == "U_Branch_Id").Value);
             AccSafeCashEditAddDTO accSafeCashDTO = await _accSafeCashRepo.GetAccSafeCashByIdAsync(id, userCode);
+            if (accSafeCashDTO.AprovedFlag == "APR")
+            {
+                return RedirectToAction("Index", "AccSafeCash", new { docType = "SFCIN", flagType = "oneANDtwo", message= "This permission is approved !" });
+
+            }
             return View(accSafeCashDTO);
         }
 
@@ -160,7 +165,10 @@ namespace ProSoft.UI.Areas.Treasury.Controllers
                 //ViewBag.ErrorMessage = "Cannot delete this record because it has related data in another table.";
                 return RedirectToAction("Index", "AccSafeCash", new { docType = "SFCIN", flagType = "oneANDtwo" , errorMessage });
             }
-
+            if (accSafeCash.AprovedFlag == "APR")
+            {
+                return RedirectToAction("Index", "AccSafeCash", new { docType = "SFCIN", flagType = "oneANDtwo", message = "This permission is approved !" });
+            }
             await _accSafeCashRepo.DeleteAsync(accSafeCash);
             await _accSafeCashRepo.SaveChangesAsync();
             return RedirectToAction("Index", "AccSafeCash",new { docType = "SFCIN", flagType = "oneANDtwo" });
